refactor(frontend): shared interaction state
Refactor shared interaction state as `lib/state/interactions`.
