+++ frontend/src/details/DeselectButton.tsx
@@ -1,6 +1,6 @@
 import { Close } from '@mui/icons-material';
 import { IconButton } from '@mui/material';
-import { selectionState } from 'state/interactions/interaction-state';
+import { selectionState } from 'lib/state/interactions/interaction-state';
 import { useResetRecoilState } from 'recoil';
 
 export const DeselectButton = ({ interactionGroup, title }) => {
