+++ frontend/src/details/regions/RegionDetails.tsx
@@ -1,6 +1,6 @@
 import { useRecoilValue } from 'recoil';
 
-import { selectionState } from 'state/interactions/interaction-state';
+import { selectionState } from 'lib/state/interactions/interaction-state';
 import { Box } from '@mui/material';
 import { RegionDetailsContent } from './RegionDetailsContent';
 import { SidePanel } from 'details/SidePanel';
