+++ frontend/src/lib/data-map/DataMapTooltip.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { useRecoilValue } from 'recoil';
 
-import { hoverPositionState } from 'state/interactions/interaction-state';
+import { hoverPositionState } from 'lib/state/interactions/interaction-state';
 
 export const DataMapTooltip: FC<{ children: React.ReactNode }> = ({ children }) => {
   const tooltipXY = useRecoilValue(hoverPositionState);
